webapp/requirements.txt
flask
flask-restful
flask-minify
flask-socketio
pymysql
pytoml
dotted
sqlalchemy
openpyxl
eventlet
kombu
pillow
numpy